RQ1/code/requirements.txt
gitdb==4.0.12
GitPython==3.1.44
numpy==2.2.4
pandas==2.2.3
psutil==7.0.0
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
smmap==5.0.2
tzdata==2025.1
